+++ src/components/sidebar/Sidebar.module.css
@@ -13,7 +13,7 @@
   flex-direction: column;
   justify-content: flex-start;
   align-items: center;
-  width: 200px;
+  min-width: 200px;
   color: white;
   box-sizing: border-box;
   padding-block: 10px;
@@ -30,34 +30,46 @@
   box-sizing: border-box;
 }
 
-.WaypointPanel {
-  overflow-x: hidden;
-  overflow-y: scroll;
-  position: absolute;
-  top: 0;
-  padding: 10px;
-  width: 300px;
-  background-color: var(--background-light-gray);
+.SidebarItem {
   color: white;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  border-radius: 0px 0px 10px 0px;
-}
-
-/* .Icon {
-  height: 48px;
-
-  border-radius: 50%;
-  width: 48px;
-  font-size: 24px;
-  text-align: center;
+  border: 2px solid transparent;
+  height: 1.5em;
+  /* background-color: var(--accent-purple); */
+  /* border-radius: 50% !important; */
+  width: 100%;
+  font-size: 18px;
   flex-shrink: 0;
-  margin-top: 0.5rem;
-  margin-bottom: 0.5rem;
+  margin-top: 0.25rem;
+  margin-bottom: 0.25rem;
   margin-left: auto;
   margin-right: auto;
   padding-top: auto;
   padding-bottom: auto;
   box-sizing: border-box;
-} */
+  color: white;
+  text-align: left;
+  display: grid;
+  grid-template-columns: 1em auto 1em;
+  align-items: center;
+  gap: 8px;
+  padding-right: 8px;
+  user-select: none;
+}
+.SidebarItem:hover {
+  background-color: var(--darker-purple);
+}
+.selected {
+  border: 2px solid var(--select-yellow);
+}
+.SidebarItem .SidebarIcon {
+  height: 1em;
+}
+.SidebarItem .SidebarLabel {
+  color: white;
+}
+.SidebarItem .SidebarRightIcon {
+  height: 1em;
+  padding-right: 8px;
+}
+
+
