+++ src/components/sidebar/SidebarRobotConfig.tsx
@@ -1,11 +1,9 @@
 import Settings from "@mui/icons-material/Settings";
 import { observer } from "mobx-react";
-import { getIdentifier } from "mobx-state-tree";
+import { safeGetIdentifier } from "../../util/mobxutils";
 import React, { Component } from "react";
-import { CSSProperties } from "styled-components";
 import DocumentManagerContext from "../../document/DocumentManager";
-import { IHolonomicWaypointStore } from "../../document/DocumentModel";
-import styles from "./SidebarWaypoint.module.css";
+import styles from "./Sidebar.module.css";
 
 type Props = {
   context: React.ContextType<typeof DocumentManagerContext>;
@@ -22,20 +20,18 @@ class SidebarWaypoint extends Component<Props, State> {
   render() {
     // apparently we have to dereference this here instead of inline in the class name
     // Otherwise the component won't rerender when it changes
-    // let selected = this.context.model.uiState.appPage== 2;
     let selected = this.context.model.robotConfig.selected;
-    // let selected = getIdentifier(this.context.model.uiState.selectedSidebarItem) == this.context.model.robotConfig.identifier;
     return (
       <div
-        className={styles.Container + (selected ? ` ${styles.selected}` : "")}
+        className={styles.SidebarItem + (selected ? ` ${styles.selected}` : "")}
         onClick={() => {
           this.context.model.uiState.setSelectedSidebarItem(
             this.context.model.robotConfig
           );
         }}
       >
-        <Settings></Settings>
-        Robot Config
+        <Settings className={styles.SidebarIcon}></Settings>
+        <span className={styles.SidebarLabel}>Robot Config</span>
       </div>
     );
   }
