--- src/components/sidebar/SidebarWaypoint.module.css
@@ -1,39 +1,0 @@
-.Container {
-  color: white;
-  border: 2px solid transparent;
-  height: 1.5em;
-  /* background-color: var(--accent-purple); */
-  /* border-radius: 50% !important; */
-  width: 100%;
-  font-size: 18px;
-  flex-shrink: 0;
-  margin-top: 0.25rem;
-  margin-bottom: 0.25rem;
-  margin-left: auto;
-  margin-right: auto;
-  padding-top: auto;
-  padding-bottom: auto;
-  box-sizing: border-box;
-  color: white;
-  text-align: left;
-  display: flex;
-  align-items: center;
-  justify-content: left;
-  gap: 8px;
-  user-select: none;
-}
-.Container:hover {
-  background-color: var(--darker-purple);
-}
-.selected {
-  border: 2px solid var(--select-yellow);
-}
-.Container .SidebarIcon {
-  height: 1em;
-}
-.Container .SidebarLabel {
-  color: white;
-}
-.Container .SidebarRightIcon {
-  height: 1em;
-}
