--- src/util/NumberEntry.module.css
@@ -1,41 +1,0 @@
-.Container {
-  position: relative;
-  background: transparent;
-  display: inline-flex;
-  flex-direction: row;
-  justify-content: center;
-  gap: 10px;
-  padding-top: 3px;
-  padding-bottom: 3px;
-  width: 100%;
-  max-width: 100%;
-}
-.Title {
-  text-align: right;
-}
-.Number {
-  width: 3rem;
-  background-color: transparent;
-  border: none;
-  color: white;
-  align-self: center;
-  flex-grow: 1;
-  flex-shrink: 1;
-  border-bottom: solid white 2px;
-}
-.Number:disabled {
-  color: transparent;
-}
-.Suffix {
-  width: 20%;
-  min-width: 4rem;
-  text-align: left;
-  flex-shrink: 0;
-}
-.Checkbox {
-  min-width: 10%;
-  margin-left: auto;
-}
-.Container.NoCheckbox .Checkbox {
-  visibility: hidden;
-}
