--- src/util/NumberEntry.tsx
@@ -1,81 +1,0 @@
-import { observer } from "mobx-react";
-import React, { Component } from "react";
-import styles from "./NumberEntry.module.css";
-
-type Props = {
-  longestTitle: string;
-  title: string;
-  suffix: string;
-  enabled: boolean;
-  number: number;
-  setNumber: (arg0: number) => void;
-  setEnabled: (arg0: boolean) => void;
-  showCheckbox?: boolean;
-};
-
-type State = {};
-class NumberEntry extends Component<Props, State> {
-  numberRef: React.RefObject<HTMLInputElement>;
-  constructor(props: Props) {
-    super(props);
-    this.setEnabled = this.setEnabled.bind(this);
-    this.setNumber = this.setNumber.bind(this);
-    this.numberRef = React.createRef<HTMLInputElement>();
-  }
-  setEnabled(event: React.ChangeEvent<HTMLInputElement>) {
-    this.props.setEnabled(event.target.checked);
-  }
-  setNumber(event: React.ChangeEvent<HTMLInputElement>) {
-    let value = event.target.value;
-    if (value === "+" || value === "-" || value === ".") return;
-    let input = Number.parseFloat(value);
-    if (!Number.isNaN(input)) {
-      this.props.setNumber(input);
-    }
-  }
-  correctNumber() {
-    if (this.numberRef.current) {
-      this.numberRef.current.value = `${this.props.number}`;
-    }
-  }
-  componentDidMount(): void {
-    this.correctNumber();
-  }
-
-  render() {
-    this.correctNumber();
-    return (
-      <>
-        <span
-          className={styles.Title}
-          style={{ width: `${this.props.longestTitle.length}ch` }}
-        >
-          {this.props.title}
-        </span>
-        <input
-          ref={this.numberRef}
-          type="text"
-          className={styles.Number}
-          disabled={!this.props.enabled}
-          onChange={this.setNumber}
-          onBlur={(e) => this.correctNumber()}
-          autoComplete="off"
-          autoCorrect="off"
-          autoCapitalize="off"
-        ></input>
-        <span className={styles.Suffix}>{this.props.suffix}</span>
-        {this.props.showCheckbox ? (
-          <input
-            type="checkbox"
-            className={styles.Checkbox}
-            checked={this.props.enabled}
-            onChange={this.setEnabled}
-          ></input>
-        ) : (
-          <span></span>
-        )}
-      </>
-    );
-  }
-}
-export default observer(NumberEntry);
